Bug fixes and feature additions
- Removed permissions related to bank info, as all bank info should be public
- Various tweaks and code cleanup
- Bank volume limits are now carried in long values for safety
- Rehashed class structure of Guis (added new abstract superclasses SinglePageGui and MultiPageGui) for better polymorphism and inheritance
- Hopefully Gui icon lore will now always be gray by default
- Fixed error in Gui generics that prevented type safety
- Added enum value GuiType.ACCOUNT_SHULKER_CONTENTS
- Cleaned up and added some assertions to BankUtils
- Tweaked message
- Reworded some permission descriptions, added notes, changed defaults

+++ src/main/java/com/monst/bankingplugin/config/Config.java
@@ -144,7 +144,7 @@ public class Config {
 	 * The default bank volume limit for players whose limit is not set via a
 	 * permission.
 	 */
-	public static int maximumBankVolume;
+	public static long maximumBankVolume;
 	
 	/**
 	 * Whether a bank owner should be allowed to create an account at their own
@@ -481,7 +481,7 @@ public class Config {
 		defaultBankLimit = config.getInt("default-limits.bank");
 		defaultAccountLimit = config.getInt("default-limits.account");
 		minimumBankVolume = Math.max(config.getInt("bank-size-limits.minimum"), 0);
-		maximumBankVolume = Math.max(config.getInt("bank-size-limits.maximum"), 0);
+		maximumBankVolume = Math.max(config.getLong("bank-size-limits.maximum"), 0);
 		allowSelfBanking = config.getBoolean("allow-self-banking");
 		confirmOnRemove = config.getBoolean("confirm-on-remove");
 		confirmOnRemoveAll = config.getBoolean("confirm-on-removeall");
